+++ ChangeLog.markdown
@@ -2,6 +2,7 @@
 
 ## 0.9.2 still in beta (grab latest stable from https://github.com/vedmack/yadcf/releases)
 
+* New filter type: date date custom, a date filter based on custom function, filter_type: "date_custom_func", https://github.com/vedmack/yadcf/issues/409
 * Fixed language url / scrollX/Y filters does not load https://github.com/vedmack/yadcf/issues/292
 * Some of additional fixed issues https://github.com/vedmack/yadcf/issues/328
 * New feature - call a callback function after yadcf init is done - onInitComplete (read docs for more info)